+++ src/navigation/nav_sensing/mcecsbot_rssi_localization/WAP.h
@@ -7,26 +7,26 @@ class WAP
   public:
     WAP();
     WAP(const string, const string);
-    WAP(const string, const string, const int, const int);
-    WAP(const string, const string, const int, const int, const int);
+    WAP(const string, const string, const float, const float);
+    WAP(const string, const string, const float, const float, const float);
     ~WAP();
     string GetAddress();
     string GetSignalLevel();
-    int GetX();
-    int GetY();
-    int GetZ();
+    float GetX();
+    float GetY();
+    float GetZ();
     void SetAddress(const string);
     void SetSignalLevel(const string);
-    void SetX(const int);
-    void SetY(const int);
-    void SetZ(const int);
-    void SetXY(const int, const int);
-    void SetXYZ(const int, const int, const int);
+    void SetX(const float);
+    void SetY(const float);
+    void SetZ(const float);
+    void SetXY(const float, const float);
+    void SetXYZ(const float, const float, const float);
 
   private:
     string _address;
     string _signalLevel;
-    int _x;
-    int _y;
-    int _z;
+    float _x;
+    float _y;
+    float _z;
 };
